+++ symja_android_library/tools/src/main/java/org/matheclipse/core/examples/Example.java
@@ -9,14 +9,13 @@ import org.matheclipse.parser.client.SyntaxError;
 import org.matheclipse.parser.client.math.MathException;
 
 /**
- * Used as an example for the Java usage in the github wiki. See:
- * <a href="https://github.com/axkr/symja_android_library/wiki/Java-usage">/wiki/Java-usage</a>
- *
+ * Used as an example for the Java usage in the github wiki. See: <a
+ * href="https://github.com/axkr/symja_android_library/wiki/Java-usage">/wiki/Java-usage</a>
  */
 public class Example {
 	public static void main(String[] args) {
 		try {
-			ExprEvaluator util = new ExprEvaluator(false, 100);
+      ExprEvaluator util = new ExprEvaluator(false, (short) 100);
 
 			// Convert an expression to the internal Java form:
 			// Note: single character identifiers are case sensitive
